refactor: de-dupe courseBlocks.factory; move to shared
