+++ src/courseware/CoursewareContainer.test.jsx
@@ -14,7 +14,7 @@ import tabMessages from '../tab-page/messages';
 import { initializeMockApp } from '../setupTest';
 
 import CoursewareContainer from './CoursewareContainer';
-import buildSimpleCourseBlocks from './data/__factories__/courseBlocks.factory';
+import { buildSimpleCourseBlocks } from '../shared/data/__factories__/courseBlocks.factory';
 import initializeStore from '../store';
 import { appendBrowserTimezoneToUrl } from '../utils';
