+++ src/courseware/data/__factories__/index.js
@@ -1,3 +1,2 @@
-import './courseBlocks.factory';
 import './courseMetadata.factory';
 import './sequenceMetadata.factory';
