+++ src/shared/data/__factories__/courseBlocks.factory.js
@@ -1,5 +1,5 @@
 import { Factory } from 'rosie'; // eslint-disable-line import/no-extraneous-dependencies
-import '../../../shared/data/__factories__/block.factory';
+import './block.factory';
 
 // Generates an Array of block IDs, either from a single block or an array of blocks.
 const getIds = (attr) => {
@@ -53,7 +53,53 @@ Factory.define('courseBlocks')
 /**
  * Builds a course with a single chapter, sequence, and unit.
  */
-export default function buildSimpleCourseBlocks(courseId, title, options = {}) {
+export function buildSimpleCourseBlocks(courseId, title, options = {}) {
+  const unitBlocks = options.unitBlocks || [Factory.build(
+    'block',
+    { type: 'vertical' },
+    { courseId },
+  )];
+  const sequenceBlock = options.sequenceBlock || [Factory.build(
+    'block',
+    { type: 'sequential', children: unitBlocks.map(block => block.id) },
+    { courseId },
+  )];
+  const sectionBlock = options.sectionBlock || Factory.build(
+    'block',
+    { type: 'chapter', children: sequenceBlock.map(block => block.id) },
+    { courseId },
+  );
+  const courseBlock = options.courseBlocks || Factory.build(
+    'block',
+    { type: 'course', display_name: title, children: [sectionBlock.id] },
+    { courseId },
+  );
+  return {
+    courseBlocks: options.courseBlocks || Factory.build(
+      'courseBlocks',
+      {
+        courseId,
+        hasScheduledContent: options.hasScheduledContent || false,
+        title: 'Demo Course',
+      },
+      {
+        units: unitBlocks,
+        sequence: sequenceBlock,
+        section: sectionBlock,
+        course: courseBlock,
+      },
+    ),
+    unitBlocks,
+    sequenceBlock,
+    sectionBlock,
+    courseBlock,
+  };
+}
+
+/**
+ * Builds a course with a single chapter and sequence, but no units.
+ */
+export function buildMinimalCourseBlocks(courseId, title, options = {}) {
   const sequenceBlock = options.sequenceBlock || [Factory.build(
     'block',
     { display_name: 'Title of Sequence', type: 'sequential' },
@@ -85,8 +131,10 @@ export default function buildSimpleCourseBlocks(courseId, title, options = {}) {
         sequence: sequenceBlock,
         section: sectionBlock,
         course: courseBlock,
+        units: [],
       },
     ),
+    unitBlocks: [],
     sequenceBlock,
     sectionBlock,
     courseBlock,
