+++ src/shared/data/__factories__/index.js
@@ -1,1 +1,2 @@
 import './block.factory';
+import './courseBlocks.factory';
